@@ -1,15 +1,17 @@
-use std::{collections::HashMap, sync::Arc, time::Duration};
+use std::{fmt::Display, sync::Arc, time::Duration};
 
+use anyhow::{Context, Result};
+use backoff::{backoff::Backoff, ExponentialBackoff};
 use futures::{FutureExt, StreamExt};
-use miette::{Context, IntoDiagnostic};
+use pallas::ledger::addresses::Address;
 use ratatui::crossterm::event::Event as CrosstermEvent;
-use tokio::sync::mpsc;
-use utxorpc::{
-    spec::sync::{any_chain_block::Chain, BlockRef, FetchBlockRequest},
-    CardanoSyncClient, TipEvent,
+use tokio::{
+    sync::{mpsc, RwLock},
+    time::sleep,
 };
+use utxorpc::{CardanoSyncClient, TipEvent};
 
-use crate::{types::DetailedBalance, wallet::types::Wallet};
+use crate::types::DetailedBalance;
 
 use super::{ChainBlock, ExplorerContext};
 
@@ -20,12 +22,30 @@ pub enum Event {
     Tick,
 }
 
+#[derive(Clone, Debug, PartialEq)]
+pub enum ConnectionState {
+    Connecting,
+    Connected,
+    Retrying,
+    Disconnected,
+}
+impl Display for ConnectionState {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ConnectionState::Connecting => write!(f, "connecting"),
+            ConnectionState::Connected => write!(f, "connected"),
+            ConnectionState::Retrying => write!(f, "retrying"),
+            ConnectionState::Disconnected => write!(f, "disconnected"),
+        }
+    }
+}
+
 #[derive(Clone, Debug)]
 pub enum AppEvent {
     Reset(u64),
     NewTip(ChainBlock),
     UndoTip(ChainBlock),
-    BalanceUpdate((String, DetailedBalance)),
+    State(ConnectionState),
 }
 
 #[derive(Debug)]
@@ -41,26 +61,31 @@ impl EventHandler {
         Self { receiver }
     }
 
-    pub async fn next(&mut self) -> miette::Result<Event> {
+    pub async fn next(&mut self) -> Result<Event> {
         self.receiver
             .recv()
             .await
-            .ok_or(miette::Report::msg("failed to receive event"))
+            .ok_or(anyhow::anyhow!("failed to receive event"))
     }
 }
 
 struct EventTask {
     sender: mpsc::UnboundedSender<Event>,
     context: Arc<ExplorerContext>,
+    state: RwLock<ConnectionState>,
 }
 
 impl EventTask {
     fn new(sender: mpsc::UnboundedSender<Event>, context: Arc<ExplorerContext>) -> Self {
-        Self { sender, context }
+        Self {
+            sender,
+            context,
+            state: RwLock::new(ConnectionState::Disconnected),
+        }
     }
 
-    async fn run(self) -> miette::Result<()> {
-        let keys = async || -> miette::Result<()> {
+    async fn run(self) -> Result<()> {
+        let keys = async || -> Result<()> {
             let mut reader = crossterm::event::EventStream::new();
 
             loop {
@@ -72,10 +97,10 @@ impl EventTask {
 
         let sender = async {
             self.sender.closed().await;
-            Ok::<_, miette::Error>(())
+            Ok::<_, anyhow::Error>(())
         };
 
-        let ticks = async || -> miette::Result<()> {
+        let ticks = async || -> Result<()> {
             let tick_rate = Duration::from_secs(1);
             let mut tick = tokio::time::interval(tick_rate);
             loop {
@@ -84,137 +109,143 @@ impl EventTask {
             }
         };
 
-        let follow_tip = async { self.follow_tip().await };
+        let follow_tip = async { self.run_follow_tip().await };
 
         tokio::try_join!(sender, keys(), follow_tip, ticks())?;
         Ok(())
     }
 
-    fn send(&self, event: Event) -> miette::Result<()> {
-        self.sender
-            .send(event)
-            .into_diagnostic()
-            .context("sending event")
+    fn send(&self, event: Event) -> Result<()> {
+        self.sender.send(event).context("sending event")
     }
 
-    async fn get_balance(&self, wallet: &Wallet) -> miette::Result<DetailedBalance> {
+    async fn update_balance(&self, address: Address, balance: DetailedBalance) {
         self.context
-            .provider
-            .get_detailed_balance(&wallet.address(self.context.provider.is_testnet()))
+            .wallets
+            .write()
             .await
+            .entry(address)
+            .and_modify(|w| w.balance = balance);
     }
 
-    async fn check_balances(
-        &self,
-        balances: &mut HashMap<String, DetailedBalance>,
-    ) -> miette::Result<()> {
-        for wallet in self.context.store.wallets() {
-            let key = wallet
-                .address(self.context.provider.is_testnet())
-                .to_string();
-            let new = self.get_balance(wallet).await?;
-            match balances.get(&key) {
-                Some(old) => {
-                    if new != *old {
-                        balances.insert(key.clone(), new.clone());
-                        self.send(Event::App(AppEvent::BalanceUpdate((key, new))))?;
-                    }
-                }
-                None => {
-                    balances.insert(key.clone(), new.clone());
-                    self.send(Event::App(AppEvent::BalanceUpdate((key, new))))?;
-                }
-            };
+    async fn get_balance(&self, address: &Address) -> Result<DetailedBalance> {
+        self.context.provider.get_detailed_balance(address).await
+    }
+
+    async fn check_balances(&self) -> Result<()> {
+        let items: Vec<(Address, DetailedBalance)> = {
+            let wallets = self.context.wallets.read().await;
+            wallets
+                .iter()
+                .map(|(addr, wallet)| (addr.clone(), wallet.balance.clone()))
+                .collect()
+        };
+
+        for (address, old_balance) in items {
+            let new_balance = self.get_balance(&address).await?;
+
+            if new_balance != old_balance {
+                self.update_balance(address.clone(), new_balance).await;
+            }
         }
 
         Ok(())
     }
 
-    async fn follow_tip(&self) -> miette::Result<()> {
-        let mut balances = HashMap::new();
-        for wallet in self.context.store.wallets() {
-            let key = wallet
-                .address(self.context.provider.is_testnet())
-                .to_string();
-            let value = self.get_balance(wallet).await?;
-            self.send(Event::App(AppEvent::BalanceUpdate((
-                key.clone(),
-                value.clone(),
-            ))))?;
-            balances.insert(key, value);
-        }
+    async fn update_connection(&self, connection: ConnectionState) -> Result<()> {
+        *self.state.write().await = connection.clone();
+        self.send(Event::App(AppEvent::State(connection)))
+    }
+
+    async fn run_follow_tip(&self) -> Result<()> {
+        self.update_connection(ConnectionState::Connecting).await?;
+
+        let max_elapsed_time = Duration::from_secs(60 * 5);
+
+        let mut backoff = ExponentialBackoff {
+            max_elapsed_time: Some(max_elapsed_time),
+            ..Default::default()
+        };
 
         loop {
-            let mut client: CardanoSyncClient = self.context.provider.client().await?;
-            let mut tip = client.follow_tip(vec![]).await.unwrap();
+            if self.follow_tip().await.is_err() {
+                if self.state.read().await.clone() == ConnectionState::Connected {
+                    backoff = ExponentialBackoff {
+                        max_elapsed_time: Some(max_elapsed_time),
+                        ..Default::default()
+                    };
+                }
 
-            while let Ok(event) = tip.event().await {
-                match event {
-                    TipEvent::Apply(block) => {
-                        let header = block.parsed.clone().unwrap().header.unwrap();
-                        let tx_count = match block.parsed {
-                            Some(parsed) => match parsed.body {
-                                Some(body) => body.tx.len(),
-                                None => 0,
-                            },
-                            None => 0,
-                        };
+                self.update_connection(ConnectionState::Retrying).await?;
 
-                        let response = client
-                            .fetch_block(FetchBlockRequest {
-                                r#ref: vec![BlockRef {
-                                    hash: header.hash.clone(),
-                                    slot: header.slot,
-                                    height: header.height,
-                                }],
-                                ..Default::default()
-                            })
-                            .await
-                            .unwrap();
-                        let fetch_block_response = response.into_inner();
-                        let body = match &fetch_block_response.block.first().unwrap().chain {
-                            Some(chain) => match chain {
-                                Chain::Cardano(block) => block.body.clone(),
-                            },
-                            None => None,
-                        };
-
-                        let chainblock = ChainBlock {
-                            slot: header.slot,
-                            hash: header.hash.to_vec(),
-                            number: header.height,
-                            tx_count,
-                            body,
-                        };
-
-                        self.send(Event::App(AppEvent::NewTip(chainblock)))?;
-                        self.check_balances(&mut balances).await?;
-                    }
-                    TipEvent::Undo(block) => {
-                        let header = block.parsed.clone().unwrap().header.unwrap();
-                        let tx_count = match block.parsed {
-                            Some(parsed) => match parsed.body {
-                                Some(body) => body.tx.len(),
-                                None => 0,
-                            },
-                            None => 0,
-                        };
-                        let chainblock = ChainBlock {
-                            slot: header.slot,
-                            hash: header.hash.to_vec(),
-                            number: header.height,
-                            tx_count,
-                            body: None,
-                        };
-                        self.send(Event::App(AppEvent::UndoTip(chainblock)))?;
-                        self.check_balances(&mut balances).await?;
-                    }
-                    TipEvent::Reset(point) => {
-                        self.send(Event::App(AppEvent::Reset(point.slot)))?;
-                        self.check_balances(&mut balances).await?;
-                    }
+                if let Some(duration) = backoff.next_backoff() {
+                    sleep(duration).await;
+                } else {
+                    self.update_connection(ConnectionState::Disconnected)
+                        .await?;
+                    break;
                 }
             }
         }
+
+        Ok(())
+    }
+
+    async fn follow_tip(&self) -> Result<()> {
+        let addresses: Vec<Address> = {
+            let wallets = self.context.wallets.read().await;
+            wallets.keys().cloned().collect()
+        };
+        for address in addresses {
+            let value = self.get_balance(&address).await?;
+            self.update_balance(address.clone(), value.clone()).await;
+        }
+
+        let mut client: CardanoSyncClient = self.context.provider.client().await?;
+        let mut tip = client.follow_tip(vec![]).await?;
+
+        self.update_connection(ConnectionState::Connected).await?;
+
+        while let Some(event) = tip.event().await? {
+            match event {
+                TipEvent::Apply(block) => {
+                    let header = block.parsed.clone().unwrap().header.unwrap();
+                    let body = block.parsed.and_then(|b| b.body);
+                    let tx_count = body.as_ref().map_or(0, |b| b.tx.len());
+
+                    let chainblock = ChainBlock {
+                        slot: header.slot,
+                        hash: header.hash.to_vec(),
+                        number: header.height,
+                        tx_count,
+                        body,
+                    };
+
+                    self.send(Event::App(AppEvent::NewTip(chainblock)))?;
+                    self.check_balances().await?;
+                }
+                TipEvent::Undo(block) => {
+                    let header = block.parsed.clone().unwrap().header.unwrap();
+                    let tx_count = block.parsed.and_then(|p| p.body).map_or(0, |b| b.tx.len());
+
+                    let chainblock = ChainBlock {
+                        slot: header.slot,
+                        hash: header.hash.to_vec(),
+                        number: header.height,
+                        tx_count,
+                        body: None,
+                    };
+
+                    self.send(Event::App(AppEvent::UndoTip(chainblock)))?;
+                    self.check_balances().await?;
+                }
+                TipEvent::Reset(point) => {
+                    self.send(Event::App(AppEvent::Reset(point.index)))?;
+                    self.check_balances().await?;
+                }
+            }
+        }
+
+        Err(anyhow::anyhow!("Tip stream ended unexpectedly"))
     }
 }
