@@ -1,7 +1,6 @@
-pub mod accounts_tab;
 pub mod activity;
-pub mod blocks_tab;
 pub mod footer;
 pub mod header;
-pub mod help;
-pub mod transactions_tab;
+pub mod popups;
+pub mod tabs;
+
