@@ -1,14 +1,15 @@
+use anyhow::bail;
 use clap::Parser;
-use miette::bail;
 
 #[derive(Parser)]
 pub struct Args {
-    /// Name of the provider to test connection with. If undefined will use default.
+    /// Name of the provider to test connection with. If undefined will use
+    /// default.
     #[arg(long)]
     name: Option<String>,
 }
 
-pub async fn run(args: Args, ctx: &crate::Context) -> miette::Result<()> {
+pub async fn run(args: Args, ctx: &crate::Context) -> anyhow::Result<()> {
     let provider = match args.name {
         Some(name) => ctx.store.find_provider(&name),
         None => ctx.store.default_provider(),
