@@ -1,11 +1,12 @@
+use anyhow::bail;
 use clap::Parser;
-use miette::bail;
 
 use crate::output::OutputFormatter;
 
 #[derive(Parser)]
 pub struct Args {
-    /// Name of the wallet to show the balance of. If undefined, will use default
+    /// Name of the wallet to show the balance of. If undefined, will use
+    /// default
     name: Option<String>,
 
     /// Name of the provider to use. If undefined, will use default
@@ -16,7 +17,7 @@ pub struct Args {
     detail: bool,
 }
 
-pub async fn run(args: Args, ctx: &crate::Context) -> miette::Result<()> {
+pub async fn run(args: Args, ctx: &crate::Context) -> anyhow::Result<()> {
     let wallet = match args.name {
         Some(name) => ctx.store.find_wallet(&name),
         None => ctx.store.default_wallet(),
